✅ Complete categories.py modularization
- Created map_locations/categories.py with 10 functions + 2 constants
- Reduced core.py from 2,552 to 2,313 lines (-239 lines net)
- Updated __init__.py to import from categories module
- All backward compatibility maintained
- Testing: ✅ All imports and functionality work correctly

Progress: 17% Complete (1/6 modules)
Next: utils.py (Priority 1: Low Risk)
