+++ map_locations/__init__.py
@@ -10,6 +10,15 @@ __version__ = "0.2.1"
 __author__ = "Lavi Shpigelman"
 __email__ = "shpigi+map_locations@gmail.com"
 
+# Import category functions
+from .categories import (
+    COLOR_CATEGORIES,
+    filter_locations_by_category,
+    get_available_categories,
+    get_location_category,
+    get_type_color,
+)
+
 # Import main functions for easy access
 from .core import (
     Location,
@@ -48,13 +57,18 @@ __all__ = [
     "get_available_types",
     "get_available_tags",
     "get_available_neighborhoods",
+    "get_available_categories",
     "filter_locations_by_type",
     "filter_locations_by_tags",
     "filter_locations_by_neighborhood",
+    "filter_locations_by_category",
     "get_location_summary",
     "validate_location_data",
     "create_sample_locations",
     "show_locations_with_google_maps",
+    "get_location_category",
+    "get_type_color",
+    "COLOR_CATEGORIES",
 ]
